--- docs/gui/modlistman.rst
@@ -1,17 +1,0 @@
-gui/modlistman
-===========
-
-.. dfhack-tool::
-    :summary: Simple modlist manager.
-    :tags: dfhack interface
-
-Adds an optional overlay to the mod list screen that
-allows you to save and load mod list presets, as well
-as set a default mod list preset for new worlds.
-
-Usage
------
-
-::
-
-    gui/modlistman
