+++ gui/mod-manager.lua
@@ -1,365 +1,467 @@
--- a graphical mod manager for df
-local gui=require 'gui'
-local widgets=require 'gui.widgets'
---[====[
+-- Simple modlist manager
+--@ module = true
 
-gui/mod-manager
-===============
-A simple way to install and remove small mods, which are not included
-in DFHack.  Examples are `available here <https://github.com/warmist/df-mini-mods>`_.
+local argparse = require('argparse')
+local overlay = require('plugins.overlay')
+local gui = require('gui')
+local widgets = require('gui.widgets')
+local repeatutil = require('repeat-util')
+local dialogs = require('gui.dialogs')
+local json = require('json')
+local utils = require('utils')
 
-.. image:: /docs/images/mod-manager.png
+local presets_file = json.open("dfhack-config/modpresets.json")
+local GLOBAL_KEY = 'modlistloader'
 
-Each mod is a lua script located in :file:`{<DF>}/mods/`, which MUST define
-the following variables:
+local function get_newregion_viewscreen()
+    local vs = dfhack.gui.getViewscreenByType(df.viewscreen_new_regionst, 0)
+    return vs
+end
 
-:name:          a name that is displayed in list
-:author:        mod author, also displayed
-:description:   a description of the mod
+local function get_modlist_fields(kind, viewscreen)
+    if kind == "available" then
+        return {
+            id = viewscreen.available_id,
+            numeric_version = viewscreen.available_numeric_version,
+            earliest_compat_numeric_version = viewscreen.available_earliest_compat_numeric_version,
+            src_dir = viewscreen.available_src_dir,
+            name = viewscreen.available_name,
+            displayed_version = viewscreen.available_displayed_version,
+            mod_header = viewscreen.available_mod_header,
+        }
+    elseif kind == "base_available" then
+        return {
+            id = viewscreen.base_available_id,
+            numeric_version = viewscreen.base_available_numeric_version,
+            earliest_compat_numeric_version = viewscreen.base_available_earliest_compat_numeric_version,
+            src_dir = viewscreen.base_available_src_dir,
+            name = viewscreen.base_available_name,
+            displayed_version = viewscreen.base_available_displayed_version,
+            mod_header = viewscreen.base_available_mod_header,
+        }
+    elseif kind == "object_load_order" then
+        return {
+            id = viewscreen.object_load_order_id,
+            numeric_version = viewscreen.object_load_order_numeric_version,
+            earliest_compat_numeric_version = viewscreen.object_load_order_earliest_compat_numeric_version,
+            src_dir = viewscreen.object_load_order_src_dir,
+            name = viewscreen.object_load_order_name,
+            displayed_version = viewscreen.object_load_order_displayed_version,
+            mod_header = viewscreen.object_load_order_mod_header,
+        }
+    else
+        error("Invalid kind: " .. kind)
+    end
+end
 
-Of course, this doesn't actually make a mod - so one or more of the
-following should also be defined:
+local function move_mod_entry(viewscreen, to, from, mod_id, mod_version)
+    local to_fields = get_modlist_fields(to, viewscreen)
+    local from_fields = get_modlist_fields(from, viewscreen)
 
-:raws_list:     a list (table) of file names that need to be copied over to df raws
-:patch_entity:  a chunk of text to patch entity
-                *TODO: add settings to which entities to add*
-:patch_init:    a chunk of lua to add to lua init
-:patch_dofile:  a list (table) of files to add to lua init as "dofile"
-:patch_files:   a table of files to patch
+    local mod_index = nil
+    for i, v in ipairs(from_fields.id) do
+        local version = from_fields.numeric_version[i]
+        if v.value == mod_id and version == mod_version then
+            mod_index = i
+            break
+        end
+    end
 
-                :filename:  a filename (in raws folder) to patch
-                :patch:     what to add
-                :after:     a string after which to insert
+    if mod_index == nil then
+        return false
+    end
 
-:guard:         a token that is used in raw files to find additions and remove them on uninstall
-:guard_init:    a token for lua file
-:[pre|post]_(un)install:
-                Callback functions, which can trigger more complicated behavior
+    for k, v in pairs(to_fields) do
+        if type(from_fields[k][mod_index]) == "userdata" then
+            v:insert('#', from_fields[k][mod_index]:new())
+        else
+            v:insert('#', from_fields[k][mod_index])
+        end
+    end
 
-]====]
-local entity_file=dfhack.getDFPath().."/raw/objects/entity_default.txt"
-local init_file=dfhack.getDFPath().."/raw/init.lua"
-local mod_dir=dfhack.getDFPath().."/hack/mods"
+    for k, v in pairs(from_fields) do
+        v:erase(mod_index)
+    end
 
-function fileExists(filename)
-    local file=io.open(filename,"rb")
-    if file==nil then
+    return true
+end
+
+local function enable_mod(viewscreen, mod_id, mod_version)
+    return move_mod_entry(viewscreen, "object_load_order", "available", mod_id, mod_version)
+end
+
+local function disable_mod(viewscreen, mod_id, mod_version)
+    return move_mod_entry(viewscreen, "available", "object_load_order", mod_id, mod_version)
+end
+
+local function get_active_modlist(viewscreen)
+    local t = {}
+    local fields = get_modlist_fields("object_load_order", viewscreen)
+    for i, v in ipairs(fields.id) do
+        local version = fields.numeric_version[i]
+        table.insert(t, { version = version, id = v.value })
+    end
+    return t
+end
+
+local function swap_modlist(viewscreen, modlist)
+    local current = get_active_modlist(viewscreen)
+    for _, v in ipairs(current) do
+        disable_mod(viewscreen, v.id, v.version)
+    end
+
+    local failures = {}
+    for _, v in ipairs(modlist) do
+        if not enable_mod(viewscreen, v.id, v.version) then
+            table.insert(failures, v.id)
+        end
+    end
+    return failures
+end
+
+ModmanageMenu = defclass(ModmanageMenu, widgets.Window)
+ModmanageMenu.ATTRS {
+    view_id = "modman_menu",
+    frame_title = "Modlist Manager",
+    frame_style = gui.WINDOW_FRAME,
+
+    resize_min = { w = 30, h = 15 },
+    frame = { w = 40, t = 10, b = 15 },
+
+    resizable = true,
+    autoarrange_subviews=false,
+}
+
+local function save_new_preset(preset_name)
+    local viewscreen = get_newregion_viewscreen()
+    local modlist = get_active_modlist(viewscreen)
+    table.insert(presets_file.data, { name = preset_name, modlist = modlist })
+    presets_file:write()
+end
+
+local function remove_preset(idx)
+    if idx > #presets_file.data then
         return
+    end
+
+    table.remove(presets_file.data, idx)
+    presets_file:write()
+end
+
+local function overwrite_preset(idx)
+    if idx > #presets_file.data then
+        return
+    end
+
+    local viewscreen = get_newregion_viewscreen()
+    local modlist = get_active_modlist(viewscreen)
+    presets_file.data[idx].modlist = modlist
+    presets_file:write()
+end
+
+local function load_preset(idx)
+    if idx > #presets_file.data then
+        return
+    end
+
+    local viewscreen = get_newregion_viewscreen()
+    local modlist = presets_file.data[idx].modlist
+    local failures = swap_modlist(viewscreen, modlist)
+
+    if #failures > 0 then
+        local failures_str = ""
+        for _, v in ipairs(failures) do
+            failures_str = failures_str .. v .. "\n"
+        end
+        dialogs.showMessage("Warning", "Failed to load some mods", COLOR_LIGHTRED)
+    end
+end
+
+local function find_preset_by_name(name)
+    for i, v in ipairs(presets_file.data) do
+        if v.name == name then
+            return i
+        end
+    end
+end
+
+local function rename_preset(idx, new_name)
+    if idx > #presets_file.data then
+        return
+    end
+
+    presets_file.data[idx].name = new_name
+    presets_file:write()
+end
+
+local function toggle_default(idx)
+    if idx > #presets_file.data then
+        return
+    end
+
+    if presets_file.data[idx].default then
+        presets_file.data[idx].default = false
+        presets_file:write()
     else
-        file:close()
-        return true
-    end
-end
-if not fileExists(init_file) then
-    local initFile=io.open(init_file,"a")
-    initFile:close()
-end
-function copyFile(from,to) --oh so primitive
-    local filefrom=io.open(from,"rb")
-    local fileto=io.open(to,"w+b")
-    local buf=filefrom:read("*a")
-    printall(buf)
-    fileto:write(buf)
-    filefrom:close()
-    fileto:close()
-end
-function patchInit(initFileName,patch_guard,code)
-    local initFile=io.open(initFileName,"a")
-    initFile:write(string.format("\n%s\n%s\n%s",patch_guard[1],
-        code,patch_guard[2]))
-    initFile:close()
-end
-function patchDofile( luaFileName,patch_guard,dofile_list,mod_path )
-    local luaFile=io.open(luaFileName,"a")
-    luaFile:write(patch_guard[1].."\n")
-    for _,v in ipairs(dofile_list) do
-        local fixed_path=mod_path:gsub("\\","/")
-        luaFile:write(string.format("dofile('%s/%s')\n",fixed_path,v))
-    end
-    luaFile:write(patch_guard[2].."\n")
-    luaFile:close()
-end
-function patchFile(file_name,patch_guard,after_string,code)
-    local input_lines=patch_guard[1].."\n"..code.."\n"..patch_guard[2]
-
-    local badchars="[%:%[%]]"
-    local find_string=after_string:gsub(badchars,"%%%1") --escape some bad chars
-
-    local entityFile=io.open(file_name,"r")
-    local buf=entityFile:read("*all")
-    entityFile:close()
-    local entityFile=io.open(file_name,"w+")
-    buf=string.gsub(buf,find_string,after_string.."\n"..input_lines)
-    entityFile:write(buf)
-    entityFile:close()
-end
-function findGuards(str,start,patch_guard)
-    local pStart=string.find(str,patch_guard[1],start)
-    if pStart==nil then return nil end
-    local pEnd=string.find(str,patch_guard[2],pStart)
-    if pEnd==nil then error("Start guard token found, but end was not found") end
-    return pStart-1,pEnd+#patch_guard[2]+1
-end
-function findGuardsFile(filename,patch_guard)
-    local file=io.open(filename,"r")
-    local buf=file:read("*all")
-    return findGuards(buf,1,patch_guard)
-end
-function unPatchFile(filename,patch_guard)
-    local file=io.open(filename,"r")
-    local buf=file:read("*all")
-    file:close()
-
-    local newBuf=""
-    local pos=1
-    local lastPos=1
-    repeat
-        local endPos
-        pos,endPos=findGuards(buf,lastPos,patch_guard)
-        newBuf=newBuf..string.sub(buf,lastPos,pos)
-        if endPos~=nil then
-            lastPos=endPos
+        for i, v in ipairs(presets_file.data) do
+            v.default = false
         end
-    until pos==nil
-
-    local file=io.open(filename,"w+")
-    file:write(newBuf)
-    file:close()
+        presets_file.data[idx].default = true
+        presets_file:write()
+    end
 end
-function checkInstalled(dfMod) --try to figure out if installed
-    if dfMod.checkInstalled then
-        return dfMod.checkInstalled()
-    else
-        if dfMod.raws_list then
-            for k,v in pairs(dfMod.raws_list) do
-                if fileExists(dfhack.getDFPath().."/raw/objects/"..v) then
-                    return true,v
-                end
+
+function ModmanageMenu:init()
+    local presetList
+
+    local function refresh_list()
+        presets_file:read()
+        local presets = utils.clone(presets_file.data, true)
+        local default_set = false
+        for _, v in ipairs(presets) do
+            v.text = v.name
+
+            if v.default and not default_set then
+                v.text = v.text .. " (default)"
+                default_set = true
             end
         end
-        if dfMod.patch_entity then
-            if findGuardsFile(entity_file,dfMod.guard)~=nil then
-                return true,"entity_default.txt"
-            end
-        end
-        if dfMod.patch_files then
-            for k,v in pairs(dfMod.patch_files) do
-                if findGuardsFile(dfhack.getDFPath().."/raw/objects/"..v.filename,dfMod.guard)~=nil then
-                    return true,"v.filename"
-                end
-            end
-        end
-        if dfMod.patch_init then
-            if findGuardsFile(init_file,dfMod.guard_init)~=nil then
-                return true,"init.lua"
-            end
-        end
+    
+        presetList:setChoices(presets)
     end
-end
-manager=defclass(manager,gui.FramedScreen)
 
-function manager:init(args)
-    self.mods={}
-    local mods=self.mods
-    local mlist=dfhack.internal.getDir(mod_dir)
-
-    if mlist==nil or #mlist==0 then
-        qerror("Mod directory not found! Are you sure it is in:"..mod_dir)
-    end
-    for k,v in ipairs(mlist) do
-        if v~="." and v~=".." then
-            local f,modData=pcall(dofile,mod_dir.."/".. v .. "/init.lua")
-            if f then
-                mods[modData.name]=modData
-                modData.guard=modData.guard or {">>"..modData.name.." patch","<<End "..modData.name.." patch"}
-                modData.guard_init={"--"..modData.guard[1],"--"..modData.guard[2]}
-                modData.path=mod_dir.."/"..v..'/'
-            end
-        end
-    end
-    ---show thingy
-    local modList={}
-    for k,v in pairs(self.mods) do
-        table.insert(modList,{text=k,data=v})
-    end
+    presetList = widgets.List {
+        frame = { b = 5 },
+        on_double_click = function(idx, current)
+            load_preset(idx)
+        end,
+    }
+    
+    refresh_list()
 
     self:addviews{
+        presetList,
 
+        widgets.HotkeyLabel{
+            frame = { l = 0, b = 1, w = 15 },
+            key = "CUSTOM_S",
+            label = "Save current",
+            on_activate = function()
+                dialogs.showInputPrompt("Enter preset name", nil, nil, "", function(t)
+                    local existing_idx = find_preset_by_name(t)
+                    if existing_idx then
+                        dialogs.showYesNoPrompt(
+                            "Confirmation", 
+                            "Overwrite " .. t .. "?",
+                            nil, 
+                            function() 
+                                overwrite_preset(existing_idx)
+                                refresh_list()
+                            end
+                        )
 
-        widgets.Panel{subviews={
-            widgets.Label{
-                text="Info:",
-                frame={t=1,l=1}
-            },
-            widgets.Label{
-                text="<no-info>",
-                --text={text=self:callback("formDescription")},
-                view_id='info',
-                frame={t=2,l=1},
-            },
-            widgets.Label{
-                text={"Author:",{text=self:callback("formAuthor")}},
-                view_id='author',
-                frame={b=5,l=1}
-            },
-            widgets.Label{
-                text={
-                {text="Install",key="CUSTOM_I",key_sep="()",disabled=self:callback("curModInstalled"),
-                    on_activate=self:callback("installCurrent")},NEWLINE,
-                {text="Uninstall",key="CUSTOM_U",key_sep="()",enabled=self:callback("curModInstalled"),
-                    on_activate=self:callback("uninstallCurrent")},NEWLINE,
-                {text="Settings",key="CUSTOM_S",key_sep="()",enabled=self:callback("hasSettings")},NEWLINE,
-                {text="Exit",key="LEAVESCREEN",key_sep="()",},NEWLINE
-                },
-                frame={l=1,b=0}
-            },
+                        return
+                    else
+                        save_new_preset(t)
+                        refresh_list()
+                    end
+                end)
+            end,
         },
-        frame={l=21,t=1,b=1}
+
+        widgets.HotkeyLabel{
+            frame = { l = 16, b = 1, w = 15 },
+            key = "CUSTOM_R",
+            label = "Rename",
+            on_activate = function()
+                local idx, current = presetList:getSelected()
+
+                if not idx then
+                    return
+                end
+
+                dialogs.showInputPrompt("Enter new name", nil, nil, current.name, function(t)
+                    local existing_idx = find_preset_by_name(t)
+                    
+                    if existing_idx then
+                        if existing_idx == idx then
+                            return
+                        end
+
+                        dialogs.showYesNoPrompt(
+                            "Confirmation", 
+                            "Overwrite " .. t .. "?",
+                            nil, 
+                            function() 
+                                remove_preset(existing_idx)
+                                rename_preset(idx, t)
+                                refresh_list()
+                            end
+                        )
+                    else
+                        rename_preset(idx, t)
+                        refresh_list()
+                    end
+                end)
+            end,
         },
-        widgets.Panel{subviews={
-            widgets.Label{
-                text="Mods:",
-                frame={t=1,l=1}
-            },
-            widgets.List{
-                choices=modList,
-                frame={t=2,l=1},
-                on_select=self:callback("selectMod")
-            },
-            },
-            frame={w=20,t=1,l=1,b=1}
+
+        widgets.HotkeyLabel{
+            frame = { l = 0, b = 2, w = 15 },
+            key = "SELECT",
+            label = "Load",
+            on_activate = function()
+                local idx, current = presetList:getSelected()
+
+                if not idx then
+                    return
+                end
+
+                load_preset(idx)
+            end,
+        },
+
+        widgets.HotkeyLabel{
+            frame = { l = 16, b = 2, w = 15 },
+            key = "CUSTOM_D",
+            label = "Delete",
+            on_activate = function()
+                local idx, current = presetList:getSelected()
+
+                if not idx then
+                    return
+                end
+
+                dialogs.showYesNoPrompt(
+                    "Confirmation", 
+                    "Delete " .. current.text .. "?", 
+                    nil, 
+                    function()
+                        remove_preset(idx)
+                        refresh_list()
+                    end
+                )
+            end,
+        },
+
+        widgets.HotkeyLabel{
+            frame = { l = 0, b = 0, w = 20 },
+            key = "CUSTOM_Q",
+            label = "Set as default",
+            on_activate = function()
+                local idx, current = presetList:getSelected()
+
+                if not idx then
+                    return
+                end
+
+                toggle_default(idx)
+                refresh_list()
+            end
         },
     }
-    self:updateState()
+end
 
-end
-function manager:postinit(args)
-    self:selectMod(1,{data=self.selected})-- workaround for first call, now the subviews are constructed
-end
-function manager:curModInstalled()
-    return self.selected.installed
-end
-function manager:hasSettings()
-    return self.selected.settings -- somehow add the entity selection as a default, if it mods entities
-end
-function manager:formDescription()
-    local ret={}
-    if self.selected.description then
-        return self.selected.description
-        --[[
-        local str=self.selected.description:split("\n")
-        for _,s in ipairs(str) do
-            table.insert(ret,{text=s})
-            table.insert(ret,NEWLINE)
-        end
-        return ret]]
-    else
-        return "<no-info>"
-    end
-end
-function manager:formAuthor()
-    return self.selected.author or "<no-info>"
-end
-function manager:selectMod(idx,choice)
-    self.selected=choice.data
-    if self.subviews.info then
-        self.subviews.info:setText(self:formDescription())
-        self:updateLayout()
-    end
-end
-function manager:updateState()
-    for k,v in pairs(self.mods) do
-        v.installed=checkInstalled(v)
-    end
-end
-function manager:installCurrent()
-    self:install(self.selected)
-end
-function manager:uninstallCurrent()
-    self:uninstall(self.selected)
-end
-function manager:install(trgMod,force)
+ModmanageScreen = defclass(ModmanageScreen, gui.ZScreen)
+ModmanageScreen.ATTRS {
+    focus_path = "modman_screen",
+    defocusable = false,
+}
 
-    if trgMod==nil then
-        qerror 'Mod does not exist'
-    end
-    if not force then
-        local isInstalled,file=checkInstalled(trgMod) -- maybe load from .installed?
-        if isInstalled then
-            qerror("Mod already installed. File:"..file)
-        end
-    end
-    print("installing:"..trgMod.name)
-    if trgMod.pre_install then
-        trgMod.pre_install(args)
-    end
-    if trgMod.raws_list then
-        for k,v in pairs(trgMod.raws_list) do
-            copyFile(trgMod.path..v,dfhack.getDFPath().."/raw/objects/"..v)
-        end
-    end
-    if trgMod.patch_entity then
-        local entity_target="[ENTITY:MOUNTAIN]" --TODO configure
-        patchFile(entity_file,trgMod.guard,entity_target,trgMod.patch_entity)
-    end
-    if trgMod.patch_files then
-        for k,v in pairs(trgMod.patch_files) do
-            patchFile(dfhack.getDFPath().."/raw/objects/"..v.filename,trgMod.guard,v.after,v.patch)
+function ModmanageScreen:init()
+    self:addviews{
+        ModmanageMenu{}
+    }
+end
+
+ModmanageOverlay = defclass(ModmanageOverlay, overlay.OverlayWidget)
+ModmanageOverlay.ATTRS {
+    frame = { w=30, h=1 },
+    default_pos = { x=5, y=-5 },
+    viewscreens = { "new_region/Mods" },
+    default_enabled=true,
+}
+
+function ModmanageOverlay:init()
+    self:addviews{
+        widgets.HotkeyLabel{
+            frame = {l=0, t=0, w = 20, h = 1},
+            key = "CUSTOM_M",
+            view_id = "modman_open",
+            label = "Mod Manager",
+            on_activate = function()
+                ModmanageScreen{}:show()
+            end,
+        },
+    }
+end
+
+NotificationOverlay = defclass(NotificationOverlay, overlay.OverlayWidget)
+NotificationOverlay.ATTRS {
+    frame = { w=60, h=1 },
+    default_pos = { x=3, y=-2 },
+    viewscreens = { "new_region" },
+    default_enabled=true,
+    focus_path = "modman_notification",
+}
+
+function NotificationOverlay:init()
+    notification_overlay_instance = self
+
+    self:addviews{
+        widgets.Label{
+            frame = { l=0, t=0, w = 60, h = 1 },
+            view_id = "lbl",
+            text = "",
+            text_pen = { fg = COLOR_GREEN, bold = true, bg = nil },
+
+        },
+    }
+end
+
+OVERLAY_WIDGETS = {
+    button = ModmanageOverlay,
+    notification = NotificationOverlay,
+}
+
+next_notification_timer_call = 0
+notification_overlay_end = 0
+function notification_timer_fn()
+    if notification_overlay_instance and #notification_overlay_instance.subviews.lbl.text > 0 then
+        if notification_overlay_end < dfhack.getTickCount() then
+            notification_overlay_instance.subviews.lbl:setText("")
         end
     end
-    if trgMod.patch_init then
-        patchInit(init_file,trgMod.guard_init,trgMod.patch_init)
-    end
-    if trgMod.patch_dofile then
-        patchDofile(init_file,trgMod.guard_init,trgMod.patch_dofile,trgMod.path)
-    end
-    trgMod.installed=true
 
-    if trgMod.post_install then
-        trgMod.post_install(self)
-    end
-    print("done")
+    dfhack.timeout(50, 'frames', notification_timer_fn)
 end
-function manager:uninstall(trgMod)
-    print("Uninstalling:"..trgMod.name)
-    if trgMod.pre_uninstall then
-        trgMod.pre_uninstall(args)
-    end
 
-    if trgMod.raws_list then
-        for k,v in pairs(trgMod.raws_list) do
-            os.remove(dfhack.getDFPath().."/raw/objects/"..v)
+notification_timer_fn()
+
+local default_applied = false
+dfhack.onStateChange[GLOBAL_KEY] = function(sc)
+    if sc == SC_VIEWSCREEN_CHANGED then
+        local vs = get_newregion_viewscreen()
+        if vs and not default_applied then
+            default_applied = true
+            for i, v in ipairs(presets_file.data) do
+                if v.default then
+                    load_preset(i)
+
+                    if notification_overlay_instance then
+                        notification_overlay_instance.subviews.lbl:setText("*** Loaded mod list '" .. v.name .. "'!")
+                        notification_overlay_end = dfhack.getTickCount() + 5000
+                    end
+                    
+                    break
+                end
+            end
+        elseif not vs then
+            default_applied = false
         end
     end
-    if trgMod.patch_entity then
-        unPatchFile(entity_file,trgMod.guard)
-    end
-    if trgMod.patch_files then
-        for k,v in pairs(trgMod.patch_files) do
-            unPatchFile(dfhack.getDFPath().."/raw/objects/"..v.filename,trgMod.guard)
-        end
-    end
-    if trgMod.patch_init or trgMod.patch_dofile then
-        unPatchFile(init_file,trgMod.guard_init)
-    end
-
-    trgMod.installed=false
-    if trgMod.post_uninstall then
-        trgMod.post_uninstall(args)
-    end
-    print("done")
 end
-function manager:onInput(keys)
 
-    if keys.LEAVESCREEN  then
-        self:dismiss()
-    else
-        self:inputToSubviews(keys)
-    end
-
+if dfhack_flags.module then
+    return
 end
-if dfhack.gui.getCurFocus()~='title' then
-    qerror("Can only be used in title screen")
-end
-local m=manager{}
-m:show()
